Refactored pagedSalariesByKarkun to queries file.
