+++ idreesia-common/server/graphql-api/hr/salary/resolvers.js
@@ -11,8 +11,7 @@ import {
   Permissions as PermissionConstants,
 } from 'meteor/idreesia-common/constants';
 import { createMonthlySalaries } from 'meteor/idreesia-common/server/business-logic/hr/create-monthly-salaries';
-import { parse } from 'query-string';
-import { get } from 'meteor/idreesia-common/utilities/lodash';
+import { getPagedSalariesByKarkun } from './queries';
 
 export default {
   SalaryType: {
@@ -35,25 +34,6 @@ export default {
   },
 
   Query: {
-    salariesByKarkun(obj, { karkunId }, { user }) {
-      if (
-        !hasOnePermission(user._id, [
-          PermissionConstants.HR_VIEW_EMPLOYEES,
-          PermissionConstants.HR_MANAGE_EMPLOYEES,
-          PermissionConstants.HR_DELETE_EMPLOYEES,
-        ])
-      ) {
-        return [];
-      }
-
-      return Salaries.find(
-        {
-          karkunId,
-        },
-        { sort: { createdAt: -1 } }
-      ).fetch();
-    },
-
     salariesByMonth(obj, { month, jobId }, { user }) {
       if (
         !hasOnePermission(user._id, [
@@ -117,36 +97,7 @@ export default {
           totalResults: 0,
         };
       }
-
-      const params = parse(queryString);
-      const pipeline = [];
-      const { pageIndex = '0', pageSize = '20', karkunId } = params;
-
-      pipeline.push({
-        $match: {
-          karkunId: { $eq: karkunId },
-        },
-      });
-
-      const countingPipeline = pipeline.concat({
-        $count: 'total',
-      });
-
-      const nPageIndex = parseInt(pageIndex, 10);
-      const nPageSize = parseInt(pageSize, 10);
-      const resultsPipeline = pipeline.concat([
-        { $sort: { createdAt: -1 } },
-        { $skip: nPageIndex * nPageSize },
-        { $limit: nPageSize },
-      ]);
-
-      const karkunSalaries = Salaries.aggregate(resultsPipeline).toArray();
-      const totalResults = Salaries.aggregate(countingPipeline).toArray();
-
-      return Promise.all([karkunSalaries, totalResults]).then(results => ({
-        salaries: results[0],
-        totalResults: get(results[1], ['0', 'total'], 0),
-      }));
+      return getPagedSalariesByKarkun(queryString);
     },
   },
